Move missionsReducers code into missionsSlice file
